optimedlab-zchatbot/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
groq==0.9.0
langchain==0.3.21
langchain-core==0.3.45
langchain-community==0.3.18
langchain-groq==0.2.4
langchain-ollama==0.2.2
ollama==0.4.7
faiss-cpu==1.9.0
numpy==1.26.4
pydantic==2.10.6
python-dotenv==1.0.1
langchain-huggingface==0.0.3
sentence-transformers==3.0.1